Remove ExceptionHandlerExtension Move ExceptionHandlerMiddleware to WebApi.Middleware Folder
