+++ src/DonateTo.Services/DonateTo.Services.csproj
@@ -5,6 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Folder Include="Extensions\" />
     <Folder Include="Services\" />
   </ItemGroup>
 
@@ -18,7 +19,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\DonateTo.ApplicationCore\DonateTo.ApplicationCore.csproj" />
-    <ProjectReference Include="..\DonateTo.Infrastructure\DonateTo.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
